requirements.txt
djangorestframework==3.14.0
Django==4.1.2
python-dotenv==0.21.0
django-cors-headers
tablib~=3.2.1
pymysql~=1.0.2
ipapy~=0.0.9.0
django-import-export==3.0.1
django-filter~=22.1